+++ spec/controllers/individual_market/insured/interactive_identity_verifications_controller_spec.rb
@@ -3,7 +3,7 @@
 require "rails_helper"
 
 if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
-  describe Insured::InteractiveIdentityVerificationsController do
+  describe IndividualMarket::Insured::InteractiveIdentityVerificationsController do
 
     after do
       allow(TimeKeeper).to receive(:date_of_record).and_call_original
